LibSys/requirements.txt
aiohttp==3.8.1
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.3.0
charset-normalizer==2.1.1
colorama==0.4.6
Django==4.2.17
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
environs==11.2.1
exceptiongroup==1.2.2
frozenlist==1.5.0
github==1.2.7
idna==3.10
iniconfig==2.0.0
marshmallow==3.23.2
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pytest==8.3.4
python-dotenv==1.0.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
yarl==1.18.3